# Conflicts:
#   .gitignore
#   backend/build.gradle
#   backend/src/main/java/ssafy/backend/afterglow/controller/RecordController.java
#   backend/src/main/java/ssafy/backend/afterglow/controller/UserController.java
#   backend/src/main/java/ssafy/backend/afterglow/domain/Record.java
#   backend/src/main/java/ssafy/backend/afterglow/domain/User.java
#   backend/src/main/java/ssafy/backend/afterglow/repository/RecordRepository.java
#   backend/src/main/java/ssafy/backend/afterglow/repository/UserRepository.java
#   backend/src/main/java/ssafy/backend/afterglow/service/RecordService.java
#   backend/src/main/java/ssafy/backend/afterglow/service/UserService.java

+++ backend/src/main/java/ssafy/backend/afterglow/repository/RecordRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import ssafy.backend.afterglow.domain.Record;
 
 @Repository
-public interface RecordRepository extends JpaRepository<Record, Integer> {
-
+public interface RecordRepository extends JpaRepository<Record, Long> {
 }
